+++ brain/python/brain.cpp
@@ -25,6 +25,8 @@
 
 #include "arrayHelpers.h"
 
+#include "docstrings.h"
+
 #include <vmmlib/vector.hpp>
 
 namespace brain
@@ -65,12 +67,19 @@ struct URItoString
 
 BOOST_PYTHON_MODULE(_brain)
 {
+#if defined BRION_USE_SPHINX && defined BRION_USE_DOGYXGEN
+    /* Only change the default Boost.Python options for documentation if we
+       are going to get docstrings from doxygen. */
+    boost::python::docstring_options doc_options(true, true, false);
+#endif
+
     boost::python::to_python_converter< servus::URI, URItoString >();
     boost::python::to_python_converter< brain::Vector3f, Vector3fToTuple >();
 
     brain::importArray();
 
-    boost::python::enum_< brain::SynapsePrefetch >( "SynapsePrefetch" )
+    boost::python::enum_< brain::SynapsePrefetch >(
+        "SynapsePrefetch", DOXY_ENUM( brain::SynapsePrefetch ))
         .value( "none", brain::SynapsePrefetch::none )
         .value( "attributes", brain::SynapsePrefetch::attributes )
         .value( "positions", brain::SynapsePrefetch::positions )