+++ brain/python/circuit.cpp
@@ -23,6 +23,7 @@
 #include "arrayHelpers.h"
 #include "helpers.h"
 #include "types.h"
+#include "docstrings.h"
 
 #include <brain/circuit.h>
 #include <brain/synapses.h>
@@ -143,7 +144,7 @@ void export_Circuit()
 {
 
 bp::class_< Circuit, boost::noncopyable, CircuitPtr >
-    circuitWrapper( "Circuit", bp::no_init );
+    circuitWrapper( "Circuit", DOXY_CLASS( brain::Circuit ), bp::no_init );
 
 bp::scope circuitScope = circuitWrapper;
 
@@ -152,29 +153,62 @@ bp::enum_< Circuit::Coordinates >( "Coordinates" )
     .value( "local", Circuit::Coordinates::local )
     ;
 
+const auto selfarg = bp::arg( "self" );
+
+// Do not modify whitespace on DOXY_FN lines
+
 circuitWrapper
-    .def( "__init__", bp::make_constructor( Circuit_initFromURI ))
-    .def( "gids", Circuit_getAllGIDs )
-    .def( "gids", Circuit_getGIDs )
-    .def( "random_gids", Circuit_getRandomTargetGIDs )
-    .def( "random_gids", Circuit_getRandomGIDs )
-    .def( "morphology_uris", Circuit_getMorphologyURIs )
-    .def( "load_morphologies", Circuit_loadMorphologies )
-    .def( "positions", Circuit_getPositions )
-    .def( "morphology_types", Circuit_getMorphologyTypes )
-    .def( "morphology_names", Circuit_getMorphologyNames )
-    .def( "electrophysiology_types", Circuit_getElectrophysiologyTypes )
-    .def( "electrophysiology_names", Circuit_getElectrophysiologyNames )
-    .def( "transforms", Circuit_getTransforms )
-    .def( "rotations", Circuit_getRotations )
-    .def( "num_neurons", &Circuit::getNumNeurons )
+    .def( "__init__", bp::make_constructor( Circuit_initFromURI ),
+          DOXY_FN( brain::Circuit::Circuit(const URI&)))
+    .def( "gids", Circuit_getAllGIDs, ( selfarg ),
+          DOXY_FN( brain::Circuit::getGIDs() const))
+    .def( "gids", Circuit_getGIDs, ( selfarg, bp::arg( "target" )),
+          DOXY_FN( brain::Circuit::getGIDs(const std::string&) const))
+    .def( "random_gids", Circuit_getRandomTargetGIDs,
+          ( selfarg, bp::arg( "fraction" ), bp::arg( "target" )),
+          DOXY_FN( brain::Circuit::getRandomGIDs(float, const std::string&) const))
+    .def( "random_gids", Circuit_getRandomGIDs,
+          ( selfarg, bp::arg( "fraction" )),
+          DOXY_FN( brain::Circuit::getRandomGIDs(float) const))
+    .def( "morphology_uris", Circuit_getMorphologyURIs,
+          ( selfarg, bp::arg( "gids" )),
+          DOXY_FN( brain::Circuit::getMorphologyURIs ))
+    .def( "load_morphologies", Circuit_loadMorphologies,
+          ( selfarg, bp::arg( "gids" ), bp::arg( "coords" )),
+          DOXY_FN( brain::Circuit::loadMorphologies ))
+    .def( "positions", Circuit_getPositions, ( selfarg, bp::arg( "gids" )),
+          DOXY_FN( brain::Circuit::getPositions ))
+    .def( "morphology_types", Circuit_getMorphologyTypes,
+          ( selfarg, bp::arg( "gids" )),
+          DOXY_FN( brain::Circuit::getMorphologyTypes ))
+    .def( "morphology_names", Circuit_getMorphologyNames, ( selfarg ),
+          DOXY_FN( brain::Circuit::getMorphologyNames ))
+    .def( "electrophysiology_types", Circuit_getElectrophysiologyTypes,
+          ( selfarg, bp::arg( "gids" )),
+          DOXY_FN( brain::Circuit::getElectrophysiologyTypes ))
+    .def( "electrophysiology_names", Circuit_getElectrophysiologyNames,
+          ( selfarg),
+          DOXY_FN( brain::Circuit::getElectrophysiologyNames ))
+    .def( "transforms", Circuit_getTransforms,
+          ( selfarg, bp::arg( "gids" )),
+          DOXY_FN( brain::Circuit::getTransforms ))
+    .def( "rotations", Circuit_getRotations,
+          ( selfarg, bp::arg( "gids" )),
+          DOXY_FN( brain::Circuit::getRotations ))
+    .def( "num_neurons", &Circuit::getNumNeurons, ( selfarg ),
+          DOXY_FN( brain::Circuit::getNumNeurons ))
     .def( "afferent_synapses", Circuit_getAfferentSynapses,
-          ( bp::arg( "gids"), bp::arg( "prefetch" ) = SynapsePrefetch::none ))
+          ( selfarg, bp::arg( "gids"),
+            bp::arg( "prefetch" ) = SynapsePrefetch::none ),
+          DOXY_FN( brain::Circuit::getAfferentSynapses ))
     .def( "efferent_synapses", Circuit_getEfferentSynapses,
-          ( bp::arg( "gids"), bp::arg( "prefetch" ) = SynapsePrefetch::none ))
+          ( selfarg, bp::arg( "gids"),
+            bp::arg( "prefetch" ) = SynapsePrefetch::none ),
+          DOXY_FN( brain::Circuit::getEfferentSynapses ))
     .def( "projected_synapses", Circuit_getProjectedSynapses,
-          ( bp::arg( "preGids"), bp::arg( "postGIDs" ),
-            bp::arg( "prefetch" ) = SynapsePrefetch::none ))
+          ( selfarg, bp::arg( "preGids"), bp::arg( "postGIDs" ),
+            bp::arg( "prefetch" ) = SynapsePrefetch::none ),
+          DOXY_FN( brain::Circuit::getProjectedSynapses ))
     ;
 }